requirements.txt
langchain
openai
PyGithub
httpx_oauth
streamlit